Lookups on btree indexes on healpixels
A predicate spoint_dwithin_healpix(order, p1, p2, radius) is introduced
that returns true when spoints p1 and p2 are at most radius apart, and
when the corresponding healpixels at this order are also at most this
radius apart:

      spoint_dwithin(p1, p2, radius) and
      healpix_dwithin("order",
                      healpix_nest("order", p1),
                      healpix_nest("order", p2),
                      radius)

The second condition has a support function that can rewrite the
condition to "pixel 1 is in the list of pixels in the circle around p2
with radius":

      healpix_nest("order", p1)
      = ANY (healpix_disc("order",
                          healpix_nest("order", p2),
                          radius))

This can efficiently be queried using a btree index with a chosen order:
healpix_nest(10, p). The chosen index order and the order used in the
query need to match.